Make splash page show on in-development site.
Also:
- Dependency updates
- Animation changes: Always fade in
