+++ .gitignore
@@ -61,3 +61,4 @@ typings/
 public
 yarn-error.log
 *.swp
+package-lock.json
