+++ package.json
@@ -10,8 +10,7 @@
   "dependencies": {
     "@material-ui/core": "^3.9.0",
     "dotenv": "^6.1.0",
-    "gatsby": "^2.0.38",
-    "gatsby-cli": "^2.4.8",
+    "gatsby": "^2.1.10",
     "gatsby-image": "^2.0.19",
     "gatsby-plugin-less": "^2.0.7",
     "gatsby-plugin-styled-jsx": "^3.0.3",
@@ -19,7 +18,7 @@
     "gatsby-transformer-sharp": "^2.1.7",
     "less": "^3.8.1",
     "prop-types": "^15.6.2",
-    "react": "^16.6.0",
+    "react": "^16.8.2",
     "react-dom": "^16.6.0",
     "styled-jsx": "^3.2.0",
     "url-search-params": "^1.1.0"
