+++ src/components/Layout.jsx
@@ -1,14 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Header from './Header';
+import Splash from './Splash';
 import styles from './Layout.module.less';
 
-const Layout = ({ children }) => (
-  <div className={styles.container}>
-    <Header />
-    {children}
-  </div>
-);
+const Layout = ({ children }) => {
+  if (process.env.GATSBY_IN_DEV === 'true') {
+    return <Splash />;
+  }
+  return (
+    <div className={styles.container}>
+      <Header />
+      {children}
+    </div>
+  );
+};
 
 Layout.propTypes = {
   children: PropTypes.element,
